function to improve linebreaks in long URLs
This breaks in weird ways, when I try to put it in to-html.py –  it doubles at the beginning and end – and I'm not sure why. It also drove me NUTS trying to get BeautifulSoup to stop escaping brackets. But at least the functions seem to work when you pass them *only* the html that you want to parse.

Help welcome!